requirements.txt
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
python-dotenv
csv